requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.2
numpy>=1.24.0
pandas>=1.5.0
yfinance>=0.2.18
scikit-learn>=1.0.2
tensorflow>=2.12.0
xgboost>=1.7.5
newsapi-python>=0.2.7
finnhub-python>=2.4.13
tweepy>=4.10.0
praw>=7.6.0
sec-api>=1.0.0
textblob>=0.17.1
nltk>=3.6.7
beautifulsoup4>=4.11.1
pandas_market_calendars>=3.2
redis>=4.3.4
apscheduler>=3.9.1
requests>=2.26.0
pytz>=2021.3
psutil>=5.9.0
torch>=2.0.1
transformers>=4.30.2
aiohttp>=3.8.5
asyncio>=3.4.3
ta>=0.10.2
joblib>=1.2.0